refactor fhir conversion to be more extensible
