+++ backend/src/main/java/gov/cdc/usds/simplereport/service/fhirConversion/FhirConversionStrategy.java
@@ -1,0 +1,9 @@
+package gov.cdc.usds.simplereport.service.fhirConversion;
+
+import java.util.Map;
+import java.util.UUID;
+import org.hl7.fhir.r4.model.Bundle;
+
+public interface FhirConversionStrategy {
+  Bundle convertRowToFhirBundle(Map<String, String> csvRow, UUID orgId);
+}
