+++ backend/src/test/java/gov/cdc/usds/simplereport/service/AzureStorageQueueFhirReportingServiceTest.java
@@ -11,7 +11,7 @@ import static org.mockito.Mockito.when;
 
 import ca.uhn.fhir.context.FhirContext;
 import com.azure.storage.queue.QueueAsyncClient;
-import gov.cdc.usds.simplereport.api.converter.FhirConverter;
+import gov.cdc.usds.simplereport.api.converter.BulkTestResultUploadFhirConverter;
 import gov.cdc.usds.simplereport.db.model.PhoneNumber;
 import gov.cdc.usds.simplereport.db.model.auxiliary.PhoneType;
 import java.util.List;
@@ -27,14 +27,15 @@ import reactor.core.publisher.MonoSink;
 @SpringBootTest
 class AzureStorageQueueFhirReportingServiceTest {
   @Autowired GitProperties gitProperties;
-  @Autowired FhirConverter fhirConverter;
+  @Autowired BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter;
 
   @Test
   void reportAsync_NonCovidOnly() {
     var context = spy(FhirContext.class);
     var client = mock(QueueAsyncClient.class);
     AzureStorageQueueFhirReportingService service =
-        new AzureStorageQueueFhirReportingService(context, client, gitProperties, fhirConverter);
+        new AzureStorageQueueFhirReportingService(
+            context, client, gitProperties, bulkTestResultUploadFhirConverter);
 
     var multiplexTestEvent = createMultiplexTestEvent();
     ReflectionTestUtils.setField(multiplexTestEvent, "internalId", UUID.randomUUID());
@@ -67,7 +68,8 @@ class AzureStorageQueueFhirReportingServiceTest {
     var context = spy(FhirContext.class);
     var client = mock(QueueAsyncClient.class);
     AzureStorageQueueFhirReportingService service =
-        new AzureStorageQueueFhirReportingService(context, client, null, fhirConverter);
+        new AzureStorageQueueFhirReportingService(
+            context, client, null, bulkTestResultUploadFhirConverter);
 
     var multiplexTestEvent = createCovidTestEvent();
     ReflectionTestUtils.setField(multiplexTestEvent.getPatient(), "internalId", UUID.randomUUID());
